+++ src/observe.js
@@ -78,17 +78,13 @@
                     createWatch( element, bindingOptions, element.id );
 
                 } else {
-                    if ( !_configuration.safeMode ) {
-                        console.error( _configuration.attributeNotValidErrorText.replace( "{{attribute_name}}", _attribute_Name_Watch_Options ) );
-                        result = false;
-                    }
+                    logError( _configuration.attributeNotValidErrorText.replace( "{{attribute_name}}", _attribute_Name_Watch_Options ) );
+                    result = false;
                 }
 
             } else {
-                if ( !_configuration.safeMode ) {
-                    console.error( _configuration.attributeNotSetErrorText.replace( "{{attribute_name}}", _attribute_Name_Watch_Options ) );
-                    result = false;
-                }
+                logError( _configuration.attributeNotSetErrorText.replace( "{{attribute_name}}", _attribute_Name_Watch_Options ) );
+                result = false;
             }
         }